# internal/reserver/model_examination.go
/*
 * Reserver Api
 *
 * Room and ambulance reservation management for Web-In-Cloud system
 *
 * API version: 1.0.0
 * Contact: xhozza@stuba.sk
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package reserver

type Examination struct {

	// Unique identifier of the examination
	Id string `json:"id"`

	// Room ID
	Room string `json:"room"`

	// ID of the department
	Department string `json:"department"`

	// ID of the doctor
	Doctor string `json:"doctor"`

	// Name of the patient
	Patient string `json:"patient"`

	// Date and time of the examination
	Datetime string `json:"datetime"`
}